# .github/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atlas Control API
  version: 0.1.0
  description: |
    HTTP API for domain intelligence, graph pivots, and discovery campaigns.
    Workers consume jobs from NATS; CT ingestion runs as a separate service.
servers:
  - url: http://localhost:8090
    description: Local docker-compose stack
paths:
...
  /ct/config:
    get:
      summary: Get CT ingestor config
      operationId: getCtConfig
      responses:
        "200":
          description: Ingestor configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CtConfig"
    put:
      summary: Update CT ingestor config
      operationId: setCtConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CtConfig"
      responses:
        "200":
          description: Config updated
  /ct/backfill:
    post:
      summary: Enable TLD-targeted CT backfill
      operationId: startCtBackfill
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CtBackfillRequest"
      responses:
        "202":
          description: Backfill enabled
...
  /campaigns:
    get:
      summary: List campaigns
      operationId: listCampaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Campaign list
    post:
      summary: Create discovery campaign
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "201":
          description: Campaign created
  /campaigns/{id}:
    get:
      summary: Get campaign
      operationId: getCampaign
      parameters:
        - $ref: "#/components/parameters/campaignId"
      responses:
        "200":
          description: Campaign details
...
  /campaigns/{id}/expand:
    post:
      summary: Approve expansion entities
      operationId: expandCampaign
      parameters:
        - $ref: "#/components/parameters/campaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entity_ids]
              properties:
                entity_ids:
                  type: array
                  items:
                    type: string
                collectors:
                  type: array
                  items:
                    type: string
      responses:
        "202":
          description: Expansion queued
components:
  parameters:
...
    campaignId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    CtConfig:
      type: object
      properties:
        target_tlds:
          type: array
          items:
            type: string
        backfill_mode:
          type: boolean
        include_readonly:
          type: boolean
        batches_per_cycle:
          type: integer
        batch_size:
          type: integer
    CtBackfillRequest:
      type: object
      properties:
        target_tlds:
          type: array
          items:
            type: string
        include_readonly:
          type: boolean
        batches_per_cycle:
          type: integer
        batch_size:
          type: integer
    CreateCampaignRequest:
      type: object
      required: [seeds]
      properties:
        seeds:
          type: array
          items:
            type: string
        collectors:
          type: array
          items:
            type: string
        depth:
          type: integer
        limits:
          type: object
          properties:
            max_depth:
              type: integer
            max_entities:
              type: integer
...